+++ lib/TransactionValidator.js
@@ -32,7 +32,8 @@ const sdk = require('postman-collection'),
     getLastSegmentURL
   } = require('./utils/textUtils'),
   {
-    GET_METHOD
+    GET_METHOD,
+    POST_METHOD
   } = require('./utils/httpUtils'),
   {
     getArrayFrom
@@ -120,7 +121,7 @@ class TransactionValidator {
             options
           ),
           requestBodyMismatches = this.validateBody(
-            requestBody,
+            requestItem.request.body,
             wsdlObject,
             operationFromWSDL,
             false,
@@ -482,6 +483,12 @@ class TransactionValidator {
       protocol = HTTP_PROTOCOL;
       return { methodName, protocol };
     }
+
+    if (itemRequest.method && itemRequest.method === POST_METHOD && methodName === '') {
+      methodName = this.getLastURLPath(itemRequest.url);
+      protocol = HTTP_PROTOCOL;
+      return { methodName, protocol };
+    }
   }
 
   /**
@@ -576,6 +583,7 @@ class TransactionValidator {
       return (
         (isIncludedOrEmpty &&
           operation.name === itemRequestProcessedInfo.name &&
+          operation.method === itemRequestProcessedInfo.method &&
           operation.protocol === itemRequestProcessedInfo.protocol) ||
         (operation.soapActionSegment !== '' &&
           operation.soapActionSegment === itemRequestProcessedInfo.soapActionSegment)
